Move integration test attmept to app/integration folder
